1315 - Moved JsonApi files to their actual location of in data-common module and removed comments in JsonApi* files saying they are temporary since aren't temporary.
