+++ data-common/src/main/java/org/cedar/onestop/data/api/JsonApiErrorResponse.java
@@ -1,15 +1,9 @@
-package org.cedar.onestop.user.common;
-
-import org.springframework.http.HttpStatus;
-
-/**
- * Temporary copy of what should eventually exist in array-iteration module.
- */
+package org.cedar.onestop.data.api;
 
 public class JsonApiErrorResponse extends JsonApiResponse {
 
   String id;
-  HttpStatus status;
+  int status;
   String code;
   String title;
   String detail;
@@ -17,7 +11,7 @@ public class JsonApiErrorResponse extends JsonApiResponse {
   ErrorSource source;
   // There are lots other optional fields that can be added if needed.
 
-  public JsonApiErrorResponse(String id, HttpStatus status, String code, String title, String detail, JsonApiMeta meta, ErrorSource source) {
+  public JsonApiErrorResponse(String id, int status, String code, String title, String detail, JsonApiMeta meta, ErrorSource source) {
     this.id = id;
     this.status = status;
     this.code = code;
@@ -31,7 +25,7 @@ public class JsonApiErrorResponse extends JsonApiResponse {
     return id;
   }
 
-  public HttpStatus getStatus() {
+  public int getStatus() {
     return status;
   }
 
@@ -64,7 +58,7 @@ public class JsonApiErrorResponse extends JsonApiResponse {
 
     String id;
     //  JsonLinks links;
-    HttpStatus status;
+    int status;
     String code;
     String title;
     String detail;
@@ -76,31 +70,61 @@ public class JsonApiErrorResponse extends JsonApiResponse {
       return this;
     }
 
-    public Builder setStatus(HttpStatus status) {
+    /**
+     * The HTTP status code applicable to this problem.
+     * @param status HttpStatus status code
+     * @return
+     */
+    public Builder setStatus(int status) {
       this.status = status;
       return this;
     }
 
+    /**
+     * An application-specific error code, expressed as a string value.
+     * @param code String application-specific error code
+     * @return
+     */
     public Builder setCode(String code) {
       this.code = code;
       return this;
     }
 
+    /**
+     * A short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization.
+     * @param title String
+     * @return
+     */
     public Builder setTitle(String title) {
       this.title = title;
       return this;
     }
 
+    /**
+     * A human-readable explanation specific to this occurrence of the problem. Like title, this field’s value can be localized.
+     * @param detail String
+     * @return
+     */
     public Builder setDetail(String detail) {
       this.detail = detail;
       return this;
     }
 
+    /**
+     * A meta object containing non-standard meta-information about the error.
+     * @param meta JsonApiMeta A meta object containing non-standard meta-information about the error.
+     * @return
+     */
     public Builder setMeta(JsonApiMeta meta) {
       this.meta = meta;
       return this;
     }
 
+    /**
+     * An object containing references to the source of the error,
+     * @param source ErrorSource An object containing references to the source of the error,
+     * @return
+     */
     public Builder setSource(ErrorSource source) {
       this.source = source;
       return this;
