+++ data-common/src/main/java/org/cedar/onestop/data/api/JsonApiMeta.java
@@ -1,5 +1,29 @@
 package org.cedar.onestop.data.api;
 
-public class JsonApiMeta {
+import java.util.HashMap;
+import java.util.Map;
 
+public class JsonApiMeta {
+  Map nonStandardMetadata;
+
+  public JsonApiMeta(Map nonStandardMetadata) {
+    this.nonStandardMetadata = nonStandardMetadata;
+  }
+
+  public Map getNonStandardMetadata() {
+    return nonStandardMetadata;
+  }
+
+  static public class Builder {
+    Map nonStandardMetadata = new HashMap();
+
+    public Builder setNonStandardMetadata(Map nonStandardMetadata) {
+      this.nonStandardMetadata = nonStandardMetadata;
+      return this;
+    }
+
+    public JsonApiMeta build() {
+      return new JsonApiMeta(this.nonStandardMetadata);
+    }
+  }
 }
