+++ user/build.gradle
@@ -10,6 +10,8 @@ configurations {
 dependencies {
   compileOnly("org.slf4j:slf4j-api:1.7.25")
 
+  implementation(project(':data-common'))
+
   implementation("org.springframework.boot:spring-boot-starter-web")
   implementation("org.springframework.boot:spring-boot-starter-security")
   implementation("org.springframework.security:spring-security-oauth2-jose")
