--- user/src/main/java/org/cedar/onestop/user/common/JsonApiResponse.java
@@ -1,14 +1,0 @@
-package org.cedar.onestop.user.common;
-
-import java.util.*;
-
-/**
- * Temporary copy of what should eventually exist in array-iteration module.
- */
-
-abstract public class JsonApiResponse {
-  List<JsonApiData> data;
-  List<JsonApiErrorResponse> errors;
-  JsonApiMeta meta;
-//  List<JsonApiLink> links; ?
-}
