+++ user/src/main/java/org/cedar/onestop/user/controller/UserController.java
@@ -9,9 +9,7 @@ import org.cedar.onestop.user.domain.OnestopUser;
 import org.cedar.onestop.user.service.ResourceNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.cedar.onestop.user.common.JsonApiData;
-import org.cedar.onestop.user.common.JsonApiResponse;
-import org.cedar.onestop.user.common.JsonApiSuccessResponse;
+import org.cedar.onestop.data.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.annotation.Secured;
